% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_index.R
\name{ARI}
\alias{ARI}
\title{ARI}
\usage{
ARI(object, name)
}
\arguments{
\item{object}{an Seurat object that contains cluster result and the inherent category}

\item{name}{the colname that contains the genes' inherent category of the cell metadata}
}
\value{
the value of an external index that evaluate the clustering result,range from -1 to 1, the larger, the better.
}
\description{
ARI
}
