% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelectionMetric.R
\name{clusterFC}
\alias{clusterFC}
\title{clusterFC}
\usage{
clusterFC(obj, gene_all_methods)
}
\arguments{
\item{obj}{an seurat object}

\item{gene_all_methods}{gene of all methods}
}
\description{
clusterFC
}
