% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_index.R
\name{DVI}
\alias{DVI}
\title{DVI}
\usage{
DVI(object, PCnumber)
}
\arguments{
\item{object}{an Seurat object that contains cluster result}

\item{PCnumber}{the number of PCs to use to calculate the distance of two cells.It is equal to the number of PCs to clustering.}
}
\value{
the value of internal index that evaluate the clustering result. The bigger, the better.
}
\description{
DVI
}
