% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelectionMetric.R
\name{gene_correlation}
\alias{gene_correlation}
\title{gene_correlation}
\usage{
gene_correlation(obj, gene_all_methods)
}
\arguments{
\item{obj}{an seurat object}

\item{gene_all_methods}{gene of all methods}
}
\description{
gene_correlation
}
