% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalGenes.R
\name{HRG_elbowplot}
\alias{HRG_elbowplot}
\title{HRG_elbowplot}
\usage{
HRG_elbowplot(obj)
}
\arguments{
\item{obj}{an Seurat object}

\item{method}{the method to define the knee point. Value can be "first" for first derivative cutoff or "curvature" for maximum curvature cutoff.}
}
\description{
Find regional gene number
}
