% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelectionMetric.R
\name{PCA10_explanation}
\alias{PCA10_explanation}
\title{the top10 PCA explanation}
\usage{
PCA10_explanation(obj, gene_all_methods)
}
\arguments{
\item{obj}{an seurat object}

\item{gene_all_methods}{gene of all methods}
}
\description{
the top10 PCA explanation
}
