% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalGenes.R
\name{RegionalGenes}
\alias{RegionalGenes}
\alias{RegionalFeatures}
\title{RegionalGenes}
\usage{
RegionalGenes(obj)
}
\arguments{
\item{obj}{an seurat object}

\item{nfeatures}{Number of features to select as output regional features}
}
\description{
Get and set regional feature information.
}
