% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyh.R
\name{wuyh}
\alias{wuyh}
\title{Add together two numbers}
\usage{
wuyh(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
add(3,2)
}
