% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalGenes.R
\name{RegionalGenes}
\alias{RegionalGenes}
\alias{RegionalFeatures}
\title{RegionalGenes}
\usage{
RegionalGenes(obj)
}
\arguments{
\item{obj}{an seurat object}
}
\description{
Get and set regional feature information.
}
